Added React router routes.
