+++ package.json
@@ -27,6 +27,7 @@
     "pg": "^7.5.0",
     "pg-promise": "^8.5.0",
     "react": "^16.5.2",
-    "react-dom": "^16.5.2"
+    "react-dom": "^16.5.2",
+    "react-router-dom": "^4.3.1"
   }
 }
